chore(crates): Correct crate dependency graph
The motivation for this change is to trim down both the cryto and
encoding crate to reframe them as providers of behaviour and core types.
Consequently the chain crate is exposing types as well as their
respective crypto and encoding trait implemtnations. An improvemnt down
the road would be to feature gate those impls.

* move address to chain crate
* move bech32 impls to chain crate
* drop bech32 impl for signing key
