+++ app/src/component/accounts/state.rs
@@ -2,12 +2,11 @@
 #![allow(dead_code)]
 
 use async_trait::async_trait;
-use pulzaar_chain::Account;
-use pulzaar_crypto::Address;
+use pulzaar_chain::{Account, Address};
 use pulzaar_encoding::{StateReadDecode, StateWriteEncode};
 
 mod state_key {
-    use pulzaar_crypto::Address;
+    use pulzaar_chain::Address;
 
     use crate::state_key::StateKey as _;
 
@@ -89,8 +88,8 @@ impl<T: StateWriteEncode + ?Sized> AccountsWrite for T {}
 mod test {
     use penumbra_storage::{StateDelta, Storage};
     use pretty_assertions::assert_eq;
-    use pulzaar_chain::Account;
-    use pulzaar_crypto::{Address, SigningKey};
+    use pulzaar_chain::{Account, Address};
+    use pulzaar_crypto::SigningKey;
     use rand::thread_rng;
     use tempfile::tempdir;
