+++ chain/src/lib.rs
@@ -2,6 +2,7 @@
 #![allow(dead_code)]
 
 mod account;
+mod address;
 mod amount;
 mod asset;
 mod chain_id;
@@ -14,6 +15,7 @@ pub mod genesis;
 pub mod input;
 
 pub use account::Account;
+pub use address::Address;
 pub use amount::Amount;
 pub use asset::Asset;
 pub use chain_id::ChainId;
