+++ crypto/src/lib.rs
@@ -1,6 +1,4 @@
-mod address;
 mod sign_bytes;
 
-pub use address::Address;
-pub use ed25519_consensus::{Error as Ed25519Error, Signature, SigningKey, VerificationKey};
+pub use ed25519_consensus::{Signature, SigningKey, VerificationKey};
 pub use sign_bytes::SignBytes;
