+++ encoding/Cargo.toml
@@ -10,8 +10,6 @@ rust-version.workspace  = true
 publish.workspace       = true
 
 [dependencies]
-pulzaar-crypto    = { path = "../crypto" }
-
 bcs               = { workspace = true }
 bech32            = { workspace = true, features = [ "std" ]}
 eyre              = { workspace = true, features = [ "auto-install" ] }
