+++ plz/Cargo.toml
@@ -12,16 +12,17 @@ name = "plz"
 path = "src/main.rs"
 
 [dependencies]
-pulzaar-config   = { path = "../config" }
-pulzaar-crypto   = { path = "../crypto" }
-pulzaar-encoding = { path = "../encoding" }
-pulzaar-node     = { path = "../node" }
+pulzaar-chain     = { path = "../chain" }
+pulzaar-config    = { path = "../config" }
+pulzaar-crypto    = { path = "../crypto" }
+pulzaar-encoding  = { path = "../encoding" }
+pulzaar-node      = { path = "../node" }
 
-dialoguer       = { workspace = true }
-eyre            = { workspace = true, features = [ "auto-install" ] }
-lexopt          = { workspace = true }
-rand            = { workspace = true, features = [ "std_rng" ] }
-tokio           = { workspace = true, features = [ "rt-multi-thread" ] }
+dialoguer         = { workspace = true }
+eyre              = { workspace = true, features = [ "auto-install" ] }
+lexopt            = { workspace = true }
+rand              = { workspace = true, features = [ "std_rng" ] }
+tokio             = { workspace = true, features = [ "rt-multi-thread" ] }
 
 [dev-dependencies]
 radicle-cli-test  = { workspace = true }
